// src/api-client/models/retailer-dashboard-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { NextPickupDto } from './next-pickup-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ReadyToDonateDto } from './ready-to-donate-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { SurplusUrgency } from './surplus-urgency';

/**
 * 
 * @export
 * @interface RetailerDashboardResponseDto
 */
export interface RetailerDashboardResponseDto {
    /**
     * Staged and awaiting collection.
     * @type {ReadyToDonateDto}
     * @memberof RetailerDashboardResponseDto
     */
    'ready': ReadyToDonateDto;
    /**
     * The next collection due, or null when nothing is scheduled.
     * @type {NextPickupDto}
     * @memberof RetailerDashboardResponseDto
     */
    'nextPickup'?: NextPickupDto | null;
    /**
     * Lots expiring within the urgency window.
     * @type {number}
     * @memberof RetailerDashboardResponseDto
     */
    'urgentCount': number;
    /**
     * The most pressing urgency present in stock.
     * @type {SurplusUrgency}
     * @memberof RetailerDashboardResponseDto
     */
    'highestUrgency': SurplusUrgency | null;
    /**
     * Donations delivered from this branch, all time.
     * @type {number}
     * @memberof RetailerDashboardResponseDto
     */
    'deliveredCount': number;
    /**
     * Lots currently in inventory.
     * @type {number}
     * @memberof RetailerDashboardResponseDto
     */
    'inventoryCount': number;
}